:recycle: Refactor hwWallet code
