+++ i18n/locales/en/common.json
@@ -269,7 +269,6 @@
   "Login Type not recognized.": "Login Type not recognized.",
   "Logout": "Logout",
   "Look at your Ledger for confirmation": "Look at your Ledger for confirmation",
-  "Look at your Ledger for confirmation of second signature": "Look at your Ledger for confirmation of second signature",
   "Look at your Trezor %s for completing the action": "Look at your Trezor %s for completing the action",
   "Mainnet": "Mainnet",
   "Manage your application": "Manage your application",
@@ -493,9 +492,7 @@
   "Unable to connect to the node": "Unable to connect to the node",
   "Unable to connect to the node, no response from the server.": "Unable to connect to the node, no response from the server.",
   "Unable to detect the communication layer with your Hardware Wallet": "Unable to detect the communication layer with your Hardware Wallet",
-  "Unable to detect the communication layer with your Ledger Nano S": "Unable to detect the communication layer with your Ledger Nano S",
   "Unable to detect the communication layer. Is ledger connected? Is Fido U2F Extension Installed?": "Unable to detect the communication layer. Is ledger connected? Is Fido U2F Extension Installed?",
-  "Unable to detect your Ledger Nano S. Be sure your device is connected and inside the Lisk App": "Unable to detect your Ledger Nano S. Be sure your device is connected and inside the Lisk App",
   "Undo": "Undo",
   "Unlock account": "Unlock account",
   "Update download finished": "Update download finished",
